Refactorisation types
