+++ components/auth/sign-in/SimpleOAuthSignIn.tsx
@@ -1,6 +1,6 @@
 import { signIn } from "next-auth/react";
 import ExternalSignIn from "./SimpleThirdPartySignIn";
-import { SupportedOAuthProvider } from "@/types/lib/auth";
+import { SupportedOAuthProvider } from "@/types/auth";
 
 type Props = {
   type: SupportedOAuthProvider,
