+++ types/auth.d.ts
