--- types/icon.d.ts
@@ -1,6 +1,0 @@
-export type IconProps = {
-  color?: string
-  height?: number
-  width?: number
-  children?: React.ReactNode
-}
